backend/requirements..txt
pandas
mysql-connector-python
python-dotenv
requests
rich
folium
matplotlib

